+++ packages/actions-global/src/hooks/use-show-notification.ts
@@ -1,25 +1,67 @@
-import type { TNotificationMetaOptions } from '@commercetools-frontend/notifications';
+import type {
+  TNotificationMetaOptions,
+  TAddNotificationAction,
+  TNotification,
+} from '@commercetools-frontend/notifications';
 import type { TShowNotification } from '../types';
 
 import { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 import { showNotification } from '../actions';
 
-// Returns a function that dispatches a notification, pre-configured to
-// a speficic notification.
-// Example:
-//   const showSuccessNotification = useShowNotification({
-//     domain: NOTIFICATION_DOMAINS.SIDE,
-//     kind: NOTIFICATION_KINDS_SIDE.success,
-//   });
-//   showSuccessNotification({ text: "All good!" });
-export default function useShowNotification<
+type TNotificationHook<Notification extends TShowNotification> = (
+  content: Notification,
+  meta?: TNotificationMetaOptions
+) => TAddNotificationAction<Notification & TNotification>;
+
+/**
+ * Dispatch a notification.
+ *
+ * @example
+ * const showSuccessNotification = useShowNotification();
+ * showSuccessNotification({
+ *   domain: NOTIFICATION_DOMAINS.SIDE,
+ *   kind: NOTIFICATION_KINDS_SIDE.success,
+ *   text: "All good!",
+ * });
+ */
+function useShowNotification<
   Notification extends TShowNotification
-  // FIXME: remove the `notificationFragment`, it makes the typing unnecessarily complex
->(notificationFragment?: Partial<Notification>) {
+>(): TNotificationHook<Notification>;
+/**
+ * Dispatch a notification.
+ *
+ * @deprecated: Avoid passing the notification options here.
+ * Instead define them in the notification function itself.
+ *
+ * @example
+ * Bad:
+ * const showSuccessNotification = useShowNotification({
+ *   domain: NOTIFICATION_DOMAINS.SIDE,
+ *   kind: NOTIFICATION_KINDS_SIDE.success,
+ * });
+ * showSuccessNotification({
+ *   text: "All good!",
+ * });
+ *
+ * Good:
+ * const showSuccessNotification = useShowNotification();
+ * showSuccessNotification({
+ *   domain: NOTIFICATION_DOMAINS.SIDE,
+ *   kind: NOTIFICATION_KINDS_SIDE.success,
+ *   text: "All good!",
+ * });
+ */
+function useShowNotification<Notification extends TShowNotification>(
+  notificationFragment: Partial<Notification>
+): TNotificationHook<Notification>;
+
+function useShowNotification<Notification extends TShowNotification>(
+  notificationFragment?: Partial<Notification>
+): TNotificationHook<Notification> {
   const dispatch = useDispatch();
   return useCallback(
-    (content: Notification, meta?: TNotificationMetaOptions) => {
+    (content, meta) => {
       const notification = showNotification<Notification>(
         { ...notificationFragment, ...content },
         meta
@@ -29,3 +71,5 @@ export default function useShowNotification<
     [dispatch, notificationFragment]
   );
 }
+
+export default useShowNotification;