+++ packages/mc-html-template/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @commercetools-frontend/mc-html-template
 
+## 21.2.1
+
+### Patch Changes
+
+- [#2471](https://github.com/commercetools/merchant-center-application-kit/pull/2471) [`dc02733d`](https://github.com/commercetools/merchant-center-application-kit/commit/dc02733dfe14ce864e6efd36c6746892170ade3d) Thanks [@renovate](https://github.com/apps/renovate)! - Update dependencies
+
+- Updated dependencies [[`dc02733d`](https://github.com/commercetools/merchant-center-application-kit/commit/dc02733dfe14ce864e6efd36c6746892170ade3d)]:
+  - @commercetools-frontend/application-config@21.2.1
+
 ## 21.1.2
 
 ### Patch Changes