+++ packages/permissions/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @commercetools-frontend/permissions
 
+## 21.2.1
+
+### Patch Changes
+
+- [#2471](https://github.com/commercetools/merchant-center-application-kit/pull/2471) [`dc02733d`](https://github.com/commercetools/merchant-center-application-kit/commit/dc02733dfe14ce864e6efd36c6746892170ade3d) Thanks [@renovate](https://github.com/apps/renovate)! - Update dependencies
+
+- Updated dependencies [[`dc02733d`](https://github.com/commercetools/merchant-center-application-kit/commit/dc02733dfe14ce864e6efd36c6746892170ade3d), [`dd5e33ea`](https://github.com/commercetools/merchant-center-application-kit/commit/dd5e33ea58b15e32a2d9aeb1d95cb4a4aec8d069)]:
+  - @commercetools-frontend/application-shell-connectors@21.2.1
+  - @commercetools-frontend/sentry@21.2.1
+
 ## 21.0.0
 
 ### Patch Changes